+++ R/plotCountsPerBroadClass-methods.R
@@ -1,7 +1,7 @@
 #' @name plotCountsPerBroadClass
 #' @author Michael Steinbaugh, Rory Kirchner
 #' @inherit bioverbs::plotCountsPerBroadClass
-#' @note Updated 2019-08-27.
+#' @note Updated 2019-09-16.
 #'
 #' @inheritParams acidroxygen::params
 #' @param ... Additional arguments.
@@ -33,7 +33,7 @@ NULL
 
 
 
-## Updated 2019-08-27.
+## Updated 2019-09-16.
 `plotCountsPerBroadClass,SummarizedExperiment` <-  # nolint
     function(
         object,
@@ -41,28 +41,32 @@ NULL
         interestingGroups = NULL,
         trans = c("identity", "log2", "log10"),
         fill,
-        countsAxisLabel = "counts",
-        title = "Counts per broad class"
+        labels = list(
+            title = "Counts per broad class",
+            subtitle = NULL,
+            sampleAxis = NULL,
+            countAxis = "counts"
+        )
     ) {
         validObject(object)
         assert(
             isScalar(assay),
-            isGGScale(fill, scale = "discrete", aes = "fill", nullOK = TRUE),
-            isString(countsAxisLabel),
-            isString(title, nullOK = TRUE)
+            isGGScale(fill, scale = "discrete", aes = "fill", nullOK = TRUE)
         )
         trans <- match.arg(trans)
+        labels <- matchLabels(
+            labels = labels,
+            choices = eval(formals()[["labels"]])
+        )
         interestingGroups(object) <-
             matchInterestingGroups(object, interestingGroups)
         interestingGroups <- interestingGroups(object)
-        if (!identical(trans, "identity")) {
-            countsAxisLabel <- paste(trans, countsAxisLabel)
-        }
         ## Melt the count matrix into long format.
         data <- melt(
             object = object,
             assay = assay,
             min = 1L,
+            minMethod = "perRow",
             trans = trans
         )
         data <- decode(data)
@@ -113,18 +117,17 @@ NULL
                 breaks = pretty_breaks(),
                 labels = prettyNum
             ) +
-            facet_wrap(facets = sym(biotypeCol), scales = "free_y") +
-            labs(
-                title = title,
-                x = NULL,
-                y = countsAxisLabel,
-                fill = paste(interestingGroups, collapse = ":\n")
-            ) +
-            theme(
-                axis.text.x = element_blank(),
-                axis.ticks.x = element_blank(),
-                axis.title.x = element_blank()
-            )
+            facet_wrap(facets = sym(biotypeCol), scales = "free_y")
+        ## Labels.
+        if (is.list(labels)) {
+            if (!identical(trans, "identity")) {
+                labels[["y"]] <- paste(trans, labels[["y"]])
+            }
+            labels[["fill"]] <- paste(interestingGroups, collapse = ":\n")
+            names(labels)[names(labels) == "sampleAxis"] <- "x"
+            names(labels)[names(labels) == "countAxis"] <- "y"
+            p <- p + do.call(what = labs, args = labels)
+        }
         ## Fill.
         if (is(fill, "ScaleDiscrete")) {
             p <- p + fill